+++ src/main/java/com/campus/campus/domain/place/application/dto/response/google/GooglePhoto.java
@@ -3,6 +3,7 @@ package com.campus.campus.domain.place.application.dto.response.google;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record GooglePhoto(
-	@JsonProperty("photo_reference") String photoReference
+	@JsonProperty("photo_reference")
+	String photoReference
 ) {
 }
