--- src/main/java/com/campus/campus/domain/place/infrastructure/google/GoogleMapConfig.java
@@ -1,10 +1,0 @@
-package com.campus.campus.domain.place.infrastructure.google;
-
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class GoogleMapConfig {
-
-	// @Value("${map.google.places.api-key}")
-	// public String apiKey;
-}
