--- src/main/java/com/campus/campus/domain/place/infrastructure/naver/NaverMapConfig.java
@@ -1,14 +1,0 @@
-package com.campus.campus.domain.place.infrastructure.naver;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class NaverMapConfig {
-
-	@Value("${map.naver.client-id}")
-	private String clientId;
-
-	@Value("${map.naver.client-secret}")
-	private String clientSecret;
-}
